python: Fix xmlescape handling
You can't call xmlescape() on every getvalue(), some of those return
data that is already XML.  So we need some way to differentiate if
xmlescape should be called.

Add an option to OpBase name xmlprocvalue, if that is set then a
getvalue() on that Op will be escaped.  Also add OpBase descendents that
automatically do this, ending in Leaf, where it makes sense.

Signed-off-by: Corey Minyard <corey@minyard.net>