clixon_beh: Fix some more transaction framework issues
Set the default children to None, not {}, so we will catch if something
tries to scan the children of a leaf.

Wrap all getvalue() and getonevalue() calls with str() so if something
is returned as an int or something like that it will work right.

Allow YangElemValueOnly to be used for non-leafs.

Signed-off-by: Corey Minyard <corey@minyard.net>