transaction_framework: Rework how element tree is built
The was the element tree was being built was clunky at best.  It was
complicated, error-prone, and hard to understand.

This change leaves the old way in place, but adds a new way to
configure the element tree.  You now have to do only one function
per element, and you add maps and leafs and the code handles setting
up everything internally.

Signed-off-by: Corey Minyard <corey@minyard.net>