Add YANG validation for element trees
The tree you create can be compared against the YANG code loaded into
clixon and errors issued if they don't match.  This isn't automatic, but
is easy to add.

This required added a "choice" element type and adding all those in and
fixes so it acts as a passthrough.

It also required adding an unimplemented element type for elements that
aren't implemented.

And it required fixing some issues in the implementations and adapting
for the new changes.

Signed-off-by: Corey Minyard <corey@minyard.net>